refactor: 로그인 세션을 관리하는 클래스 분리
