+++ src/main/java/roomescape/config/WebConfig.java
@@ -1,15 +1,21 @@
 package roomescape.config;
 
 import java.util.List;
+import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import roomescape.domain.LoginSession;
+import roomescape.resolver.LoginInfoArgumentResolver;
 
 @Configuration
+@RequiredArgsConstructor
 public class WebConfig implements WebMvcConfigurer {
+
+    private final LoginSession loginSession;
 
     @Override
     public void addArgumentResolvers(final List<HandlerMethodArgumentResolver> resolvers) {
-        resolvers.add(new LoginInfoArgumentResolver());
+        resolvers.add(new LoginInfoArgumentResolver(loginSession));
     }
 }
