+++ src/main/java/roomescape/controller/AuthController.java
@@ -7,9 +7,9 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.config.LoginRequired;
 import roomescape.dto.LoginInfo;
 import roomescape.dto.request.AuthRequest;
+import roomescape.resolver.LoginRequired;
 import roomescape.service.AuthService;
 
 @RestController
