+++ src/main/java/roomescape/resolver/LoginRequired.java
@@ -1,4 +1,4 @@
-package roomescape.config;
+package roomescape.resolver;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
