+++ src/main/java/roomescape/service/AuthService.java
@@ -3,10 +3,10 @@ package roomescape.service;
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import roomescape.domain.LoginSession;
 import roomescape.domain.Member;
 import roomescape.dto.LoginInfo;
 import roomescape.dto.request.AuthRequest;
-import roomescape.error.AccessDeniedException;
 import roomescape.error.AuthenticationException;
 import roomescape.repository.MemberRepository;
 
@@ -14,25 +14,22 @@ import roomescape.repository.MemberRepository;
 @RequiredArgsConstructor
 public class AuthService {
 
-    private static final String LOGIN_INFO_KEY = "loginInfo";
-
     private final MemberRepository memberRepository;
+    private final LoginSession loginSession;
 
     public void login(final AuthRequest request, final HttpSession session) {
-        final Member member = memberRepository.findByEmail(request.email()).orElseThrow(
-                () -> new AuthenticationException("존재하지 않는 이메일입니다."));
+        final Member member = memberRepository.findByEmail(request.email())
+                .orElseThrow(() -> new AuthenticationException("존재하지 않는 이메일입니다."));
+
         if (!member.getPassword().equals(request.password())) {
             throw new AuthenticationException("비밀번호가 일치하지 않습니다.");
         }
 
         final LoginInfo loginInfo = new LoginInfo(member.getId(), member.getName());
-        session.setAttribute(LOGIN_INFO_KEY, loginInfo);
+        loginSession.setLoginInfo(session, loginInfo);
     }
 
     public void logout(final HttpSession session) {
-        if (session.getAttribute(LOGIN_INFO_KEY) == null) {
-            throw new AccessDeniedException("로그인 정보가 없습니다.");
-        }
-        session.removeAttribute(LOGIN_INFO_KEY);
+        loginSession.clear(session);
     }
 }
